+++ main_test.go
@@ -3,51 +3,12 @@ package main
 import (
 	"net/http"
 	"net/http/httptest"
-	"strconv"
 	"strings"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
-
-var cafeList = map[string][]string{
-	"moscow": []string{"Мир кофе", "Сладкоежка", "Кофе и завтраки", "Сытый студент"},
-}
-
-func mainHandle(w http.ResponseWriter, req *http.Request) {
-	countStr := req.URL.Query().Get("count")
-	if countStr == "" {
-		w.WriteHeader(http.StatusBadRequest)
-		w.Write([]byte("count missing"))
-		return
-	}
-
-	count, err := strconv.Atoi(countStr)
-	if err != nil {
-		w.WriteHeader(http.StatusBadRequest)
-		w.Write([]byte("wrong count value"))
-		return
-	}
-
-	city := req.URL.Query().Get("city")
-
-	cafe, ok := cafeList[city]
-	if !ok {
-		w.WriteHeader(http.StatusBadRequest)
-		w.Write([]byte("wrong city value"))
-		return
-	}
-
-	if count > len(cafe) {
-		count = len(cafe)
-	}
-
-	answer := strings.Join(cafe[:count], ",")
-
-	w.WriteHeader(http.StatusOK)
-	w.Write([]byte(answer))
-}
 
 func TestMainHandlerWhenRequireIsTrue(t *testing.T) {
 
@@ -57,10 +18,16 @@ func TestMainHandlerWhenRequireIsTrue(t *testing.T) {
 	handler := http.HandlerFunc(mainHandle)
 	handler.ServeHTTP(responseRecorder, req)
 
+	if status := responseRecorder.Code; status != http.StatusOK {
+		t.Fatalf("expected status code: %d, got %d", http.StatusOK, status)
+	}
+
 	body := responseRecorder.Body.String()
+
 	// здесь нужно добавить необходимые проверки
+
 	assert.Greater(t, len(body), 0)
-	require.Equal(t, 200, responseRecorder.Code)
+	require.Equal(t, http.StatusOK, responseRecorder.Code)
 }
 func TestMainHandlerWhenNoSuchCity(t *testing.T) {
 
@@ -71,10 +38,14 @@ func TestMainHandlerWhenNoSuchCity(t *testing.T) {
 	handler := http.HandlerFunc(mainHandle)
 	handler.ServeHTTP(responseRecorder, req)
 
+	//if status := responseRecorder.Code; status != http.StatusOK {
+	//t.Fatalf("expected status code: %d, got %d", http.StatusOK, status)
+	//}
+
 	body := responseRecorder.Body.String()
 
 	assert.Equal(t, expected, body)
-	require.Equal(t, 400, responseRecorder.Code)
+	require.Equal(t, http.StatusBadRequest, responseRecorder.Code)
 }
 func TestMainHandlerWhenCountMoreThanTotal(t *testing.T) {
 	expectation := 4
@@ -83,6 +54,10 @@ func TestMainHandlerWhenCountMoreThanTotal(t *testing.T) {
 	responseRecorder := httptest.NewRecorder()
 	handler := http.HandlerFunc(mainHandle)
 	handler.ServeHTTP(responseRecorder, req)
+
+	if status := responseRecorder.Code; status != http.StatusOK {
+		t.Fatalf("expected status code: %d, got %d", http.StatusOK, status)
+	}
 
 	body := strings.Split(responseRecorder.Body.String(), ",")
